Refactor: PostDetailComponent 좋아요 아이콘 컴포넌트로 분리
PostDetailLikeIcon 제작하여 PostDetailComponent에 적용
